+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 0.2 (2023-01-XX)
 
 - have a progress bar for long-running commands when possible
+- use [gspread](https://docs.gspread.org/) to handle the low-level GoogleAPI
+- use [gspread-dataframe](https://gspread-dataframe.readthedocs.io/) to handle converting a worksheet into a dataframe
 
 ## Version 0.1.1 (2023-01-16)